+++ cypress/integration/draft.spec.ts
@@ -19,14 +19,15 @@ describe("Draft operations", function () {
     cy.get("[data-testid=title]").type("Test title")
 
     // Save a draft
-    cy.formActions("Save as Draft")
+    cy.formActions("Save as draft")
     cy.get("div[role=alert]", { timeout: 10000 }).contains("Draft saved with")
     cy.get("[data-testid='policy-objects-list']").find("li").should("have.length", 1)
 
     // Save another draft
-    cy.formActions("New form")
+    cy.get("button").contains("Add policy").click()
+    cy.get("[data-testid=title]").should("have.value", "")
     cy.get("[data-testid=title]").type("Test title 2")
-    cy.formActions("Save as Draft")
+    cy.formActions("Save as draft")
     cy.get("div[role=alert]", { timeout: 10000 }).contains("Draft saved with")
     cy.get("[data-testid='policy-objects-list']").find("li").should("have.length", 2)
 
@@ -35,23 +36,14 @@ describe("Draft operations", function () {
     cy.get("[data-testid='policy-objects-list']").find("a").first().click()
     cy.get("h2").contains("Would you like to save draft version of this form")
     cy.get("div[role=dialog]").contains("Save").click()
+    cy.get("div[role=dialog]", { timeout: 10000 }).should("not.exist")
     cy.get("[data-testid='policy-objects-list']").find("li").should("have.length", 2)
 
-    // Delete a draft
-    cy.get("[data-testid='Draft-objects']")
-      .find("li")
-      .first()
-      .within(() => {
-        cy.get("[data-testid='Delete submission']").first().click()
-      })
-    cy.get("[data-testid='Draft-objects']").find("li", { timeout: 10000 }).should("have.length", 1)
+    // Continue first draft
+    cy.get("[data-testid='draft-policy-list-item']", { timeout: 10000 }).first().click({ force: true })
 
-    // Continue draft
-    cy.continueLatestDraft(ObjectTypes.policy)
-
-    cy.get("[data-testid=title]")
-    // Clear
-    cy.formActions("Clear form")
+    // Clear form
+    cy.optionsActions("Clear form")
     // Fill
     cy.get("[data-testid=title]").should("have.value", "")
     cy.get("[data-testid=title]").type("New title")
@@ -62,38 +54,33 @@ describe("Draft operations", function () {
     cy.get("div[role=alert]", { timeout: 10000 }).contains("Draft updated with")
 
     // Create a new form and save as draft
-    cy.formActions("New form")
+    cy.get("button").contains("Add DAC").click()
     cy.get("[data-testid=title]").should("contain.text", "")
     cy.get("[data-testid=title]").type("New title 2")
     cy.get("[data-testid=title]").should("have.value", "New title 2")
-    cy.get("select[data-testid='dacRef.accessionId']").select(1)
-    cy.formActions("Save as Draft")
+    cy.formActions("Save as draft")
 
     cy.get("div[role=alert]", { timeout: 10000 }).contains("Draft saved with")
 
     // Check that there are 2 drafts saved in drafts list
-    cy.get("ul[data-testid='Draft-objects']").find("li").should("have.length", 2)
+    cy.get("ul[data-testid='dac-objects-list']").find("li").should("have.length", 2)
 
     // Submit first form draft
     cy.continueLatestDraft(ObjectTypes.policy)
+    cy.get("select[data-testid='dacRef.accessionId']").select(1)
     cy.get("select[data-testid='policy']").select("Policy Text")
     cy.get("textarea[data-testid='policy.policyText']").type("Test policy text")
-    cy.formActions("Submit")
+    cy.formActions("Mark as ready")
     cy.get("div[role=alert]", { timeout: 10000 }).contains("Submitted with")
-    cy.get("[data-testid='Form-objects']").find("li").should("have.length", 1)
 
     // Submit second form draft
-    cy.get(`[data-testid='draft-${ObjectTypes.policy}-list-item']`).should("have.length", 1)
     // Re-query instead of continueLatestDraft -command since original query results in detached DOM element
-    cy.get(`[data-testid='draft-${ObjectTypes.policy}-list-item']`).click()
-    cy.get("select[data-testid='policy']").select("Policy Text")
+    cy.get(`[data-testid='draft-${ObjectTypes.policy}-list-item']`).first().click({ force: true })
+    cy.get("select[data-testid='policy']", { timeout: 10000 }).should("be.visible").select("Policy Text")
     cy.get("textarea[data-testid='policy.policyText']").type("Test policy text")
-    cy.formActions("Submit")
+    cy.formActions("Mark as ready")
     // Check that there are 2 submitted objects
-    cy.get("[data-testid='Form-objects']", { timeout: 10000 }).find("li").should("have.length", 2)
-
-    // Drafts list should be unmounted
-    cy.get("[data-testid='Draft-objects']").should("have.length", 0)
+    cy.get("[data-testid='policy-objects-list']", { timeout: 10000 }).find("li").should("have.length", 2)
   })
 })