+++ src/components/ErrorPageContainer.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from "react"
 
 import ErrorOutlineOutlinedIcon from "@mui/icons-material/ErrorOutlineOutlined"
-import { CustomTheme } from "@mui/material"
+import { Theme } from "@mui/material"
 import Alert from "@mui/material/Alert"
 import Avatar from "@mui/material/Avatar"
 import Card from "@mui/material/Card"
@@ -22,7 +22,7 @@ type ErrorTypeProps = {
   errorType: string
 }
 
-const useStyles = makeStyles((theme: CustomTheme) => ({
+const useStyles = makeStyles((theme: Theme) => ({
   errorContainer: {
     width: "100%",
     marginTop: 10,