+++ src/components/SubmissionWizard/WizardComponents/WizardStepper.tsx
@@ -15,6 +15,7 @@ import WizardStep from "./WizardStep"
 import { resetObjectType } from "features/wizardObjectTypeSlice"
 import { resetStepObject, updateStep } from "features/wizardStepObjectSlice"
 import { useAppDispatch, useAppSelector } from "hooks"
+import type { FormRef } from "types"
 
 // Top & bottom borders for first and last disabled elements
 const AccordionWrapper = styled("div")(({ theme }) => ({
@@ -72,7 +73,7 @@ const AccordionDetails = styled(MuiAccordionDetails)(({ theme }) => ({
  * If createSubmissionForm is passed as reference it is used to trigger correct form when clicking next.
  */
 
-const WizardStepper = () => {
+const WizardStepper = ({ formRef }: { formRef?: FormRef }) => {
   const objectTypesArray = useAppSelector(state => state.objectTypesArray)
   const submission = useAppSelector(state => state.submission)
   const currentStepObject = useAppSelector(state => state.stepObject)
@@ -141,7 +142,12 @@ const WizardStepper = () => {
             </AccordionSummary>
             {step.stepItems && step.actionButtonText && (
               <AccordionDetails>
-                <WizardStep step={stepNumber} stepItems={step.stepItems} actionButtonText={step.actionButtonText} />
+                <WizardStep
+                  step={stepNumber}
+                  stepItems={step.stepItems}
+                  actionButtonText={step.actionButtonText}
+                  formRef={formRef}
+                />
               </AccordionDetails>
             )}
           </Accordion>