+++ src/types/index.ts
@@ -58,6 +58,7 @@ export type ObjectInsideSubmission = {
 export type ObjectTags = {
   submissionType?: string
   fileName?: string
+  fileSize?: string
   displayTitle?: string
 }
 
@@ -98,8 +99,7 @@ export type SubmissionDataFromForm = {
 }
 
 export type SubmissionFolder = SubmissionDetailsWithId & { doiInfo: DoiFormDetails }
-
-export type CreateSubmissionFormRef = { current: HTMLElement | null } | undefined
+export type FormRef = { current: HTMLElement | null } | undefined
 
 export type StatusDetails = {
   status: string | null