+++ src/utils/index.tsx
@@ -10,8 +10,6 @@ export const getObjectDisplayTitle = (objectType: string, cleanedValues: ObjectD
   switch (objectType) {
     case ObjectTypes.study:
       return cleanedValues.descriptor?.studyTitle || ""
-    case ObjectTypes.dac:
-      return cleanedValues.contacts?.find(contact => contact.mainContact)?.name || ""
     default:
       return cleanedValues.title || ""
   }
@@ -20,13 +18,7 @@ export const getObjectDisplayTitle = (objectType: string, cleanedValues: ObjectD
 // Get Primary text for displaying item's title
 export const getItemPrimaryText = (item: ObjectInsideSubmissionWithTags): string => {
   if (item.tags?.displayTitle) {
-    switch (item.schema) {
-      case ObjectTypes.dac:
-      case `draft-${ObjectTypes.dac}`:
-        return `Main Contact: ${item.tags.displayTitle}`
-      default:
-        return item.tags.displayTitle
-    }
+    return item.tags.displayTitle
   } else if (item.tags?.fileName) {
     return item.tags.fileName
   }